test(app): split create flow coverage
