+++ packages/app/src/web/app-ready-terminal-screen.tsx
@@ -7,9 +7,9 @@ import type { ReadyLayoutProps } from "./app-ready-layout.js"
 import { Box, Text } from "./elements.js"
 import { TaskPanel } from "./panel-tasks.js"
 import { TerminalPanel } from "./panel-terminal.js"
-import type { TerminalExitInfo } from "./terminal-panel-runtime-types.js"
 import { type BrowserScreen, projectPickerScreen } from "./screen.js"
 import { shouldShowTerminalTabs } from "./terminal-mobile-layout.js"
+import type { TerminalExitInfo } from "./terminal-panel-runtime-types.js"
 import { terminalSessionId } from "./terminal-state.js"
 import { type ActiveTerminalSession, isPendingActiveTerminalSession, terminalTitleById } from "./terminal.js"
