+++ packages/app/tests/docker-git/app-ready-create-settings.test.ts
@@ -1,0 +1,271 @@
+import { beforeEach, describe, expect, it, vi } from "vitest"
+
+import type { submitCreateInputs } from "../../src/web/actions-projects.js"
+import { handleCreateKey, submitCreateView } from "../../src/web/app-ready-create.js"
+import {
+  type CreateFlowView,
+  createInitialFlowView,
+  createKeyEvent,
+  createSetCreateViewSpy,
+  createSettingsFlowView,
+  createSettingsFlowViewAtStep,
+  type CreateStep,
+  expectCreateArrowHandling,
+  expectCreateSideArrowBufferHandling,
+  expectCreateViewReset,
+  requireCreateViewValue,
+  requireSubmittedCreateInputs,
+  resolveCreateDisplaySteps,
+  resolveCreateFlowSteps,
+  validGithubStatus
+} from "./app-ready-create-fixture.js"
+import { makeBrowserActionContext } from "./browser-action-context-fixture.js"
+
+const mocks = vi.hoisted(() => ({
+  submitCreateInputsMock: vi.fn<typeof submitCreateInputs>()
+}))
+
+vi.mock("../../src/web/actions-projects.js", () => ({
+  submitCreateInputs: mocks.submitCreateInputsMock
+}))
+
+const submitCreateInputsMock = mocks.submitCreateInputsMock
+
+describe("app-ready-create settings", () => {
+  beforeEach(() => {
+    submitCreateInputsMock.mockReset()
+  })
+
+  it("moves between settings with arrows and clears the uncommitted buffer", () => {
+    expectCreateArrowHandling(handleCreateKey, "ArrowDown", (view) => view.step + 1)
+  })
+
+  it("wraps settings selection upward with ArrowUp and clears the uncommitted buffer", () => {
+    expectCreateArrowHandling(handleCreateKey, "ArrowUp", (view) => resolveCreateFlowSteps(view.values).length - 1)
+  })
+
+  it("fills discrete settings buffers with side arrows without applying values", () => {
+    const cases: ReadonlyArray<{
+      readonly expectedBuffer: string
+      readonly key: "ArrowLeft" | "ArrowRight"
+      readonly stepName: CreateStep
+    }> = [
+      { expectedBuffer: "none", key: "ArrowLeft", stepName: "gpu" },
+      { expectedBuffer: "all", key: "ArrowRight", stepName: "gpu" },
+      { expectedBuffer: "n", key: "ArrowLeft", stepName: "runUp" },
+      { expectedBuffer: "y", key: "ArrowRight", stepName: "runUp" },
+      { expectedBuffer: "n", key: "ArrowLeft", stepName: "mcpPlaywright" },
+      { expectedBuffer: "y", key: "ArrowRight", stepName: "mcpPlaywright" },
+      { expectedBuffer: "n", key: "ArrowLeft", stepName: "force" },
+      { expectedBuffer: "y", key: "ArrowRight", stepName: "force" }
+    ]
+
+    for (const { expectedBuffer, key, stepName } of cases) {
+      submitCreateInputsMock.mockReset()
+      expectCreateSideArrowBufferHandling(handleCreateKey, submitCreateInputsMock, key, stepName, expectedBuffer)
+    }
+  })
+
+  it("applies a side-arrow choice only after Enter", () => {
+    const { context } = makeBrowserActionContext({ githubStatus: validGithubStatus })
+    const { setCreateView, spy: setCreateViewSpy } = createSetCreateViewSpy()
+    const createView = createSettingsFlowViewAtStep("gpu", "typed")
+    const arrowEvent = createKeyEvent("ArrowRight")
+
+    const arrowHandled = handleCreateKey(arrowEvent, {
+      context,
+      controllerCwd: "/workspace",
+      createView,
+      projectsRoot: "/home/dev/.docker-git",
+      setCreateView
+    })
+    const arrowView = requireCreateViewValue(setCreateViewSpy.mock.calls[0]?.[0])
+    const enterEvent = createKeyEvent("Enter")
+
+    const enterHandled = handleCreateKey(enterEvent, {
+      context,
+      controllerCwd: "/workspace",
+      createView: arrowView,
+      projectsRoot: "/home/dev/.docker-git",
+      setCreateView
+    })
+    const enteredView = requireCreateViewValue(setCreateViewSpy.mock.calls[1]?.[0])
+
+    expect(arrowHandled).toBe(true)
+    expect(arrowView.values.gpu).toBeUndefined()
+    expect(enterHandled).toBe(true)
+    expect(enteredView.values.gpu).toBe("all")
+    expect(enteredView.step).toBe(resolveCreateDisplaySteps().indexOf("gpu"))
+    expect(enteredView.buffer).toBe("")
+    expect(submitCreateInputsMock).not.toHaveBeenCalled()
+  })
+
+  it("keeps an applied settings row selected and visible instead of submitting", () => {
+    const { context } = makeBrowserActionContext({ githubStatus: validGithubStatus })
+    const { setCreateView, spy: setCreateViewSpy } = createSetCreateViewSpy()
+    const createView: CreateFlowView = {
+      ...createSettingsFlowViewAtStep("force", "y"),
+      values: {
+        ...createSettingsFlowView().values,
+        cpuLimit: "40%",
+        enableMcpPlaywright: true,
+        gpu: "all",
+        ramLimit: "8g",
+        runUp: false
+      }
+    }
+    const event = createKeyEvent("Enter")
+
+    const handled = handleCreateKey(event, {
+      context,
+      controllerCwd: "/workspace",
+      createView,
+      projectsRoot: "/home/dev/.docker-git",
+      setCreateView
+    })
+    const enteredView = requireCreateViewValue(setCreateViewSpy.mock.calls[0]?.[0])
+
+    expect(handled).toBe(true)
+    expect(enteredView.values.force).toBe(true)
+    expect(enteredView.step).toBe(resolveCreateDisplaySteps().indexOf("force"))
+    expect(enteredView.buffer).toBe("")
+    expect(submitCreateInputsMock).not.toHaveBeenCalled()
+  })
+
+  it("navigates to the next visible row after applying a settings row", () => {
+    const { context } = makeBrowserActionContext({ githubStatus: validGithubStatus })
+    const { setCreateView, spy: setCreateViewSpy } = createSetCreateViewSpy()
+    const createView = createSettingsFlowViewAtStep("mcpPlaywright", "y")
+    const enterEvent = createKeyEvent("Enter")
+
+    handleCreateKey(enterEvent, {
+      context,
+      controllerCwd: "/workspace",
+      createView,
+      projectsRoot: "/home/dev/.docker-git",
+      setCreateView
+    })
+    const enteredView = requireCreateViewValue(setCreateViewSpy.mock.calls[0]?.[0])
+    const downEvent = createKeyEvent("ArrowDown")
+
+    const handled = handleCreateKey(downEvent, {
+      context,
+      controllerCwd: "/workspace",
+      createView: enteredView,
+      projectsRoot: "/home/dev/.docker-git",
+      setCreateView
+    })
+    const downView = requireCreateViewValue(setCreateViewSpy.mock.calls[1]?.[0])
+
+    expect(handled).toBe(true)
+    expect(downView.step).toBe(resolveCreateDisplaySteps().indexOf("force"))
+    expect(downView.values.enableMcpPlaywright).toBe(true)
+    expect(downView.buffer).toBe("")
+  })
+
+  it("clears an unconfirmed preview when navigating away from a settings row", () => {
+    const { context } = makeBrowserActionContext({ githubStatus: validGithubStatus })
+    const { setCreateView, spy: setCreateViewSpy } = createSetCreateViewSpy()
+    const createView = createSettingsFlowViewAtStep("mcpPlaywright", "y")
+    const event = createKeyEvent("ArrowDown")
+
+    const handled = handleCreateKey(event, {
+      context,
+      controllerCwd: "/workspace",
+      createView,
+      projectsRoot: "/home/dev/.docker-git",
+      setCreateView
+    })
+    const nextView = requireCreateViewValue(setCreateViewSpy.mock.calls[0]?.[0])
+
+    expect(handled).toBe(true)
+    expect(nextView.step).toBe(resolveCreateDisplaySteps().indexOf("force"))
+    expect(nextView.values.enableMcpPlaywright).toBeUndefined()
+    expect(nextView.buffer).toBe("")
+  })
+
+  it("submits settings Done with a valid active preview applied first", () => {
+    const { context } = makeBrowserActionContext({ githubStatus: validGithubStatus })
+    const { setCreateView, spy: setCreateViewSpy } = createSetCreateViewSpy()
+
+    submitCreateView({
+      context,
+      controllerCwd: "/workspace",
+      createView: createSettingsFlowViewAtStep("mcpPlaywright", "y"),
+      projectsRoot: "/home/dev/.docker-git",
+      quickCreate: false,
+      setCreateView
+    })
+
+    expect(submitCreateInputsMock).toHaveBeenCalledTimes(1)
+    expect(requireSubmittedCreateInputs(submitCreateInputsMock).enableMcpPlaywright).toBe(true)
+    expectCreateViewReset(setCreateViewSpy)
+  })
+
+  it("shows a parse error when settings Done has an invalid active preview", () => {
+    const { context } = makeBrowserActionContext({ githubStatus: validGithubStatus })
+    const { setCreateView, spy: setCreateViewSpy } = createSetCreateViewSpy()
+
+    submitCreateView({
+      context,
+      controllerCwd: "/workspace",
+      createView: createSettingsFlowViewAtStep("gpu", "bogus"),
+      projectsRoot: "/home/dev/.docker-git",
+      quickCreate: false,
+      setCreateView
+    })
+
+    expect(submitCreateInputsMock).not.toHaveBeenCalled()
+    expect(setCreateViewSpy).not.toHaveBeenCalled()
+    expect(context.setMessage).toHaveBeenCalledWith("Invalid option create: gpu must be one of: none, all, yes, no")
+  })
+
+  it("ignores settings arrows before the Settings flow starts", () => {
+    const { context } = makeBrowserActionContext({ githubStatus: validGithubStatus })
+    const { setCreateView, spy: setCreateViewSpy } = createSetCreateViewSpy()
+    const event = createKeyEvent("ArrowDown")
+
+    const handled = handleCreateKey(event, {
+      context,
+      controllerCwd: "/workspace",
+      createView: createInitialFlowView("https://github.com/org/repo"),
+      projectsRoot: "/home/dev/.docker-git",
+      setCreateView
+    })
+
+    expect(handled).toBe(false)
+    expect(event.preventDefault).not.toHaveBeenCalled()
+    expect(setCreateViewSpy).not.toHaveBeenCalled()
+    expect(context.setMessage).not.toHaveBeenCalled()
+  })
+
+  it("ignores side arrows before Settings and on free-text settings", () => {
+    const keys: ReadonlyArray<"ArrowLeft" | "ArrowRight"> = ["ArrowLeft", "ArrowRight"]
+    const views: ReadonlyArray<CreateFlowView> = [
+      createInitialFlowView("https://github.com/org/repo"),
+      createSettingsFlowViewAtStep("cpuLimit"),
+      createSettingsFlowViewAtStep("ramLimit")
+    ]
+
+    for (const key of keys) {
+      for (const createView of views) {
+        const { context } = makeBrowserActionContext({ githubStatus: validGithubStatus })
+        const { setCreateView, spy: setCreateViewSpy } = createSetCreateViewSpy()
+        const event = createKeyEvent(key)
+
+        const handled = handleCreateKey(event, {
+          context,
+          controllerCwd: "/workspace",
+          createView,
+          projectsRoot: "/home/dev/.docker-git",
+          setCreateView
+        })
+
+        expect(handled).toBe(false)
+        expect(event.preventDefault).not.toHaveBeenCalled()
+        expect(setCreateViewSpy).not.toHaveBeenCalled()
+        expect(context.setMessage).not.toHaveBeenCalled()
+      }
+    }
+  })
+})
